crypto-bot/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cryptobot"
version = "0.1.0"
description = "Modular crypto algo-trading bot (backtest, paper, live)."
authors = [{ name = "You" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "ccxt>=4.3.0",
  "pandas>=2.2.2",
  "numpy>=2.1.0",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "loguru>=0.7.2",
  "aiohttp>=3.10.5",
]

[project.scripts]
cryptobot-backtest = "cryptobot.engine.run_backtest:main"
cryptobot-paper = "cryptobot.engine.run_paper:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.5",
  "mypy>=1.11.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
